combine sender and requester
