+++ pkg/bindings/http/option.go
@@ -9,5 +9,3 @@ func WithTranscoder(factory binding.TransformerFactory) SenderOptionFunc {
 		sender.transformerFactories = append(sender.transformerFactories, factory)
 	}
 }
-
-type RequesterOptionFunc func(requester *Requester)
